% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/random_generators.R
\name{ccol}
\alias{ccol}
\title{Generate constant column}
\usage{
ccol(value, cols, size)
}
\arguments{
\item{value}{Constant value}

\item{cols}{Column name for the generated data}

\item{size}{Number of rows to be generated}
}
\description{
Repeat a constant value for a column
}

