% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/density_estimates.R
\name{epdfSample}
\alias{epdfSample}
\title{Draw a random sample from a data frame}
\usage{
epdfSample(ctable, size, count_col = "N")
}
\arguments{
\item{ctable}{Count table as returned by \code{\link{getCounts}}}

\item{size}{The number of rows}

\item{count_col}{The column name for counts, usually "N".}
}
\description{
Generate a random sample from an empirical probability density.
}
\details{
The 'cdf' and count_col columns will be discarded.
}

