% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data_exploring.R
\name{get_colnames}
\alias{get_colnames}
\title{Get table names}
\usage{
get_colnames(tbl)
}
\arguments{
\item{tbl}{Table name to search}
}
\description{
Get table column names as a character vector
}

