% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data_exploring.R
\name{get_missing}
\alias{get_missing}
\title{Fraction of missing values}
\usage{
get_missing(tbl, cols, sample_max = options("synergetr_sample_max")[[1]])
}
\arguments{
\item{tbl}{Table name to search}

\item{cols}{Column name}

\item{sample_max}{Limit the number of rows when estimating
       the fraction of missing values}
}
\description{
Compute the fraction of missing values in a colum
}

