% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data_exploring.R
\name{get_sample}
\alias{get_sample}
\title{Get sample from a table}
\usage{
get_sample(tbl, size = 1000)
}
\arguments{
\item{tbl}{Table name to search}

\item{size}{Sample size}
}
\description{
Draw a random sample from a table. If size == Inf, return all rows.
}

