% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/density_estimates.R
\name{getCounts}
\alias{getCounts}
\title{Compute Frequencies for Categorical Variables}
\usage{
getCounts(tbl, cols, compute_cdf = TRUE, min_count = 1, default = "''",
  con = options("synergetr_con")[[1]],
  sample_max = options("synergetr_sample_max")[[1]])
}
\arguments{
\item{tbl}{Table name to inspect}

\item{cols}{Vector of table fields}

\item{compute_cdf}{Add cumulative distribution function to result
set as a 'cdf' column.}

\item{min_count}{The minimum amount of times a distinct value must
appear in raw data to be included in the frequency count
(this parameter is used to exclude rare values from
appearing at all).}

\item{sample_max}{The maximum number of rows to use}
}
\description{
Count, group and arrange categorical values from the most to the
least frequent. Optionally, compute cumulative distribution
function (cdf) and suppress entries that occur less than min_count
of times.
}
\details{
If option("synergetr_con") points to a database
connection, the computation of the frequencies will be done at the
database and the \code{tbl} should be a character string
(e.g. tbl == "schemaname.table_name"). If "synergetr_con" is
not set (i.e.  equals NULL), the computations will be done in R
memory using the data.table package, and \code{tbl} can be
either the actual data.frame or its variable name as a character
string.
}

