% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/density_estimates.R
\name{getDensity}
\alias{getDensity}
\title{Compute probability density estimate from numerical fields}
\usage{
getDensity(tbl, cols, con = options("synergetr_con")[[1]], trim = 0.001,
  adj = 1, discrete = FALSE, n = 500,
  sample_max = options("synergetr_sample_max")[[1]], is_date = FALSE)
}
\arguments{
\item{tbl}{Table name}

\item{cols}{Column name}

\item{con}{Connection object to the database, or NULL for in-memory
data.frames}

\item{trim}{The quentile of data to be discarded from both ends
(min and max) for robust estimation of the probability
density}

\item{adj}{Kernel width parameter to be passed to
\code{\link[stats]{density}}}

\item{discrete}{TRUE for integer-valued output}

\item{n}{Default number of point for the kernel density estimation}

\item{sample_max}{Maximum number of data points to read from the
source table. Can be used to limit the data transfer from a
database (and the consecutive memory consumption and
computational costs)}

\item{is_date}{Convert data first into Date before casting it into
       integers (where day 0 equals 1970-01-01)}
}
\description{
One-dimensional numerical kernel density estimation based on
\code{\link[stats]{density}}.  This function abstacts away database
operations and provides sensible defauts for trimming out outlier
data, handling computational cost and handling special formats
(like Date fields).
}

