% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/density_plots.R
\name{plotCatDist}
\alias{plotCatDist}
\title{Plot Value Distribution}
\usage{
plotCatDist(tbl, cols, hlp = 0.5, ntop = 10, ...)
}
\arguments{
\item{tbl}{Table name to inspect}

\item{cols}{Vector of table fields}

\item{hlp}{Per cent of points to hightlight in graph}

\item{ntop}{Number of rows to return as a sample}
}
\description{
A combined categorical distribution computation and plot.
}
\details{
The function calls getCounts for frequency computations
which, in turn, may call database engine for data.
}

