% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/random_generators.R
\name{rchar}
\alias{rchar}
\title{Generate random character fields}
\usage{
rchar(slength = 8, cols, size, type = "0Aa")
}
\arguments{
\item{slength}{String length}

\item{cols}{Column name for the generated data}

\item{size}{Number of rows to be generated}

\item{type}{Fied type specification. "0" generates numbers, "A"
capital letters and "a" small letters. The values "0Aa"
can be combined and given in any order. The value "m"
produces a special hour-minute timestamp of the
form [00-59][00-59].}
}
\description{
Generate artificial code fields from random characters and
integers.
}
\examples{
rchar(4, "code", 20, type="A0")
}

