% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/random_generators.R
\name{rcode}
\alias{rcode}
\title{Generate random code field}
\usage{
rcode(nlevels = 10, cols, size, codevec = NULL, along = NULL, ...)
}
\arguments{
\item{nlevels}{Number of levels for the code keys}

\item{cols}{Column name for the generated data}

\item{size}{Number of rows to be generated}

\item{codevec}{Explicit vector of codes to use}

\item{along}{A vector (or one-column data.frame) to match the code
keys with.}
}
\description{
Build random code fields using \code{\link{rchar}} as the
underlying workhorse.
}
\details{
For extra parameters such as code field length and type,
see \code{\link{rchar}}.
}
\examples{
# Basic use
rcode(2, "codes", 10)
rcode(NA, "codes", 10, codevec=c("foo", "bar","xyzzy"))

# Using the along feature
docname <- sample(c("Jofn A Doc", "Jane D Doc"), 10, replace=TRUE)
rcode(along=docname, cols="doccode", type="A0", slength=3)
}

