% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/density_estimates.R
\name{rdate}
\alias{rdate}
\title{Build random dates}
\usage{
rdate(tbl, cols, size, sparse = FALSE, trim = 0.001, adj = 0.1,
  n = 1000)
}
\arguments{
\item{tbl}{Table name}

\item{cols}{Column name}

\item{size}{Number of rows for the output}

\item{sparse}{Consider and estimate also the fraction of missing data}

\item{trim}{Passed to \link{getDensity}}

\item{adj}{Passed to \link{getDensity}}

\item{n}{Passed to \link{getDensity}}
}
\description{
Based on the dates at the given table column, generate a
kernel density estimate and return a random column of dates (with
the same distribution)
}
\details{
The date distribution is estimated as a one-dimensional
kernel density with \code{\link[stats]{density}} with the standard
notation that day zero equals 1970-01-01 00:00:00 (see the output
of \code{unclass(as.Date("1970-01-01"))}), and the consecutive days
are numbered sequentially as 1,2,3...
}

