% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/density_estimates.R
\name{rfactor}
\alias{rfactor}
\title{Construct synthetic factorial data}
\usage{
rfactor(tbl, cols, size, min_count = 5, partial = FALSE, nsplits = 1,
  split_sep = " ", ...)
}
\description{
Construct synthetic factorial data
}

