% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/density_estimates.R
\name{rhetu}
\alias{rhetu}
\title{Generate Random Finnish Personal IDs}
\usage{
rhetu(tbl, cols, size, ...)
}
\arguments{
\item{tbl}{Table name}

\item{cols}{Column name}

\item{size}{Number of rows for the output}
}
\description{
Generate random Pseudo-Finnish personal IDs (hetu-codes) using
\code{\link{rpid}} function. In addition to generating nonsense,
the functions estimates the real distribution of birth years and
uses that information as a 'yearString' argument to
\code{\link{rpid}}.
}

