% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/random_generators.R
\name{rpid}
\alias{rpid}
\title{Generate Imaginary Personal IDs}
\usage{
rpid(yearString = NA, size = length(yearString))
}
\arguments{
\item{yearString}{String to contruct the year part (in Finnish
convention). See the examples.}

\item{size}{Number of elements to generate}
}
\description{
The generated codes resemble Finnish national personal id
convention in the field length. that the year of birth can be deduced from
the code in a similar way, and the field width is the same.
Otherwise, there are no similarities or well-formedness rules.
}
\examples{
rpid(NA, 10)
rpid(yearString=c("78-","03A","98-"))
}

