% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/random_generators.R
\name{rpk}
\alias{rpk}
\title{Generate random primary integer keys}
\usage{
rpk(cols, size, min_value = 1)
}
\arguments{
\item{cols}{Column name for the generated data}

\item{size}{Number of rows to be generated}

\item{min_value}{Integer value to start the sequence}
}
\description{
A convenience function to generate sequences of numbers (similar to
primary keys in database tables).
}
\examples{
rpk("id", 10)
}

