% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/density_estimates.R
\name{rtime}
\alias{rtime}
\title{Build random timestamps}
\usage{
rtime(tbl, cols, size, onlyDates = FALSE, sparse = FALSE, trim = 0.001,
  adj = 0.1, n = 1e+05, ...)
}
\arguments{
\item{tbl}{Table name}

\item{cols}{Column name}

\item{size}{Number of rows for the output}

\item{onlyDates}{Build a timestamp field that contains factually
only dates (e.g. 2017-08-02 00:00:00)}

\item{sparse}{Consider and estimate also the fraction of missing data}

\item{trim}{Passed to \link{getDensity}}

\item{adj}{Passed to \link{getDensity}}

\item{n}{Passed to \link{getDensity}}
}
\description{
Based on the timestamps at the given table column, generate a
kernel density estimate and return a random column of timestamps (with
the same distribution)
}

