\name{synergetr-package}
\alias{synergetr-package}
\alias{synergetr}
\docType{package}
\title{
  \packageTitle{synergetr}
}
\description{
  \packageDescription{synergetr}
}
\details{

  The DESCRIPTION file:
  \packageDESCRIPTION{synergetr}
  \packageIndices{synergetr}
  ~~ An overview of how to use the package, including the most important ~~
  ~~ functions ~~
}
\author{
  \packageAuthor{synergetr}

  Maintainer: \packageMaintainer{synergetr}
}
\references{
  ~~ Literature or other references for background information ~~
}


\keyword{ package }
\seealso{
  ~~ Optional links to other man pages, e.g. ~~
  ~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
